llvm/scheduler: Move compiled scheduler to its own file
